Moving `FileFromDialog` helpers into helpers/utils
